Revert "Top Tabs for iPad (#1946) r=sleroux (#1957)"
This reverts commit fa155c405b2b4335815f21d1652daa1f6b900bd9.
