+++ Utils/AppConstants.swift
@@ -180,21 +180,4 @@ public struct AppConstants {
             return true
         #endif
     }()
-    
-    ///  Enables/disables the top tabs for iPad
-    public static let MOZ_TOP_TABS: Bool = {
-        #if MOZ_CHANNEL_RELEASE
-            return true
-        #elseif MOZ_CHANNEL_BETA
-            return true
-        #elseif MOZ_CHANNEL_NIGHTLY
-            return true
-        #elseif MOZ_CHANNEL_FENNEC
-            return true
-        #elseif MOZ_CHANNEL_AURORA
-            return true
-        #else
-            return true
-        #endif
-    }()
 }
